eh: Opens windows explorer to the current directory
npp: Opens a new, empty instance of Notepad++. If a directory path is
provided, it will open all the files in that directory. If an extension
is provided, it will only open files with that extension.

helpers/SET_FILES_LIST: Collects file names from directory and sets the
collected names to the "FILES_LIST" env variable. Used by "npp" but
might be useful elsewhere (hence "helpers" dir).